requirements.txt
flask
flask-wtf
flask-sqlalchemy
flask-login
email-validator
flask_uploads